added button to customerMainActivity to go to cartActivity

+++ app/src/main/java/com/example/b07_final_project/Item.java
@@ -8,7 +8,7 @@ public class Item {
     public Item() {
     }
 
-    public Item(String name, String brand, int price) {
+    public Item(String name, String brand, double price) {
         this.name = name;
         this.brand = brand;
         this.price = price;
